working asteroids and player sprites

+++ player.py
@@ -1,11 +1,10 @@
 from pygame.math import Vector2
 from circleshape import CircleShape
 from constants import PLAYER_RADIUS, PLAYER_TURN_SPEED, PLAYER_SPEED
-from typing import Tuple
 import pygame
 
 class Player(CircleShape):
-    def __init__(self, x: int, y: int) -> None:
+    def __init__(self, x: float, y: float) -> None:
         super().__init__(x,y, PLAYER_RADIUS)
 
         self.rotation: float = 0
